+++ src/core/texteditor.h
@@ -40,7 +40,6 @@ class TextEditor : public QPlainTextEdit
   QString currentFile;
   void lineNumberAreaPaintEvent(QPaintEvent *event);
   int lineNumberAreaWidth();
-
   void openFile(const QString &fileName);
 
  protected:
@@ -72,7 +71,6 @@ class TextEditor : public QPlainTextEdit
   KSyntaxHighlighting::Repository m_repository;
   KSyntaxHighlighting::SyntaxHighlighter *m_highlighter;
   QWidget *lineNumberArea;
-  QColor m_lineNumbersColor;
 
  signals:
   void changeTitle();