+++ src/ui/preferenceswindow.cpp
@@ -32,11 +32,22 @@ void PreferencesWindow::resetAllValues(const bool isFirst)
 {
   if (isFirst) {
     ui->themeCombo->addItems(QStyleFactory::keys());
-    ui->colorCombo->addItem("Light");
-    ui->colorCombo->addItem("Dark");
+
+    /*
+     * Another implementation (better but harder) :
+     for (const auto &theme : m_repository.themes()) addItem(...);
+     * There is currently no plan to customize the theme of highlighting
+     * So this implementation is temporarily used
+    */
+
+    ui->highlightThemeCombo->addItem("Breeze Dark");
+    ui->highlightThemeCombo->addItem("Default");
+    ui->highlightThemeCombo->addItem("Printing");
+    ui->highlightThemeCombo->addItem("Solarized Dark");
+    ui->highlightThemeCombo->addItem("Solarized Light");
   }
   ui->themeCombo->setCurrentText(configManager->getStyle());
   ui->fontComboBox->setCurrentFont(QFont(configManager->getEditorFontFamily()));
   ui->spinBox->setValue(configManager->getEditorFontSize());
-  ui->colorCombo->setCurrentText(configManager->getColorTheme());
+  ui->highlightThemeCombo->setCurrentText(configManager->getColorTheme());
 }